Refs #21295: Create ParticipantBuiltinTopicData in rtps namespace
